requirements.txt
requests==2.25.1
loguru==0.5.3
translatepy==2.2
getkey==0.6.5
pyTelegramBotAPI==4.14.0
python-dotenv==0.20.0
